+++ src/lib.rs
@@ -1,0 +1,4 @@
+pub mod config;
+pub mod keyboard_hook;
+pub mod layout;
+pub mod virtual_keyboard;
